refactored code to remove dynamic routing
created new discover page component
